13_Projects/Realtime_Voice_AI_Agent_with_RAG/Codes/rag_voice_ai_agent-deployment_live/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "pipecat-ai[deepgram,elevenlabs,groq,local-smart-turn-v3,protobuf,silero-vad]>=0.0.100",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "pypdf>=6.6.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
